server/docker/mqtt-subs/collector/requirements.txt
certifi==2018.4.16
chardet==3.0.4
docker==3.3.0
docker-pycreds==0.2.3
geoip2==2.8.0
idna==2.6
maxminddb==1.3.0
paho-mqtt==1.3.1
pymongo==3.6.1
requests==2.18.4
six==1.11.0
urllib3==1.22
websocket-client==0.47.0
